[Refactor] class 및 package 이름 변경: timeline → feed
