+++ src/main/java/com/example/kp3coutsourcingproject/timeline/service/FeedService.java
@@ -1,9 +1,8 @@
-package com.example.kp3coutsourcingproject.redis.service;
+package com.example.kp3coutsourcingproject.timeline.service;
 
-import com.example.kp3coutsourcingproject.post.dto.PostResponseDto;
 import com.example.kp3coutsourcingproject.post.entity.Post;
 import com.example.kp3coutsourcingproject.post.repository.PostRepository;
-import com.example.kp3coutsourcingproject.redis.dto.RedisPostDto;
+import com.example.kp3coutsourcingproject.timeline.dto.FeedPostDto;
 import com.example.kp3coutsourcingproject.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.ListOperations;
@@ -11,32 +10,31 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.LinkedList;
 import java.util.List;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class RedisTimelineService {
-    private final RedisTemplate<String, RedisPostDto> redisTemplate;
+public class FeedService {
+    private final RedisTemplate<String, FeedPostDto> redisTemplate;
     private final PostRepository postRepository;
 
     // 가장 최근에 작성한 게시글 저장하기
     public void saveMyLastCreatedPost(User user) {
-        ListOperations<String, RedisPostDto> listOperations = redisTemplate.opsForList();
+        ListOperations<String, FeedPostDto> listOperations = redisTemplate.opsForList();
         // id가 가장 마지막인 post를 찾는다
         Post lastCreatedPost = postRepository.findTopByOrderByIdDesc().orElseThrow( () ->
                 new IllegalArgumentException("작성한 게시글이 없습니다.")
         );
 
         String key = user.getId().toString();
-        RedisPostDto value = new RedisPostDto(lastCreatedPost);
+        FeedPostDto value = new FeedPostDto(lastCreatedPost);
         listOperations.leftPush(key, value); // 저장
     }
     
     // 타임라인 게시글 조회
-    public List<RedisPostDto> getTimelinePosts(User user) {
-        ListOperations<String, RedisPostDto> listOperations = redisTemplate.opsForList();
+    public List<FeedPostDto> getTimelinePosts(User user) {
+        ListOperations<String, FeedPostDto> listOperations = redisTemplate.opsForList();
 
         String key = user.getId().toString();
         long size = listOperations.size(key) == null ? 0 : listOperations.size(key);
